# docs/swagger.yaml
definitions:
  controlador.request:
    properties:
      codigo:
        type: string
      color:
        type: string
      fechaDeCreacion:
        type: string
      id:
        type: integer
      nombre:
        type: string
      precio:
        type: number
      publicado:
        type: boolean
      stock:
        type: integer
    required:
    - codigo
    - color
    - fechaDeCreacion
    - nombre
    - precio
    - publicado
    - stock
    type: object
  web.Response:
    properties:
      code:
        type: string
      data: {}
      error:
        type: string
    type: object
info:
  contact:
    name: API Support
    url: https://developers.mercadolibre.com.ar/es_ar/terminos-y-condiciones
  description: This API handles MELI productos
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: MELI Bootcamp API
  version: "1.0"
paths:
  /productos:
    get:
      consumes:
      - application/json
      description: get products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: List products
      tags:
      - Products
    post:
      consumes:
      - application/json
      description: store products
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Product to store
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controlador.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Store products
      tags:
      - Products
  /productos/id:
    delete:
      consumes:
      - application/json
      description: delete product by id
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Delete product
      tags:
      - Products
    get:
      consumes:
      - application/json
      description: get product by id
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Get product by ID
      tags:
      - Products
    patch:
      consumes:
      - application/json
      description: updates name and price of product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: Name and price value to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controlador.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Updates Name and Price of product
      tags:
      - Products
    put:
      consumes:
      - application/json
      description: updates a product
      parameters:
      - description: token
        in: header
        name: token
        required: true
        type: string
      - description: product to update
        in: body
        name: product
        required: true
        schema:
          $ref: '#/definitions/controlador.request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web.Response'
      summary: Updates a product
      tags:
      - Products
swagger: "2.0"